requirements.txt
pytest~=6.2.5
setuptools~=67.6.0
pytest-xdist~=3.2.1